+++ src/main/java/roomescape/client/payment/PaymentRepository.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.client.payment;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.Optional;
