+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -9,8 +9,8 @@ import roomescape.auth.annotation.LoginMemberId;
 import roomescape.exception.PaymentConfirmException;
 import roomescape.member.dto.MemberResponse;
 import roomescape.member.service.MemberService;
-import roomescape.payment.Payment;
-import roomescape.payment.PaymentService;
+import roomescape.client.payment.Payment;
+import roomescape.client.payment.service.PaymentService;
 import roomescape.registration.domain.reservation.dto.ReservationDto;
 import roomescape.registration.domain.reservation.service.ReservationService;
 import roomescape.registration.domain.waiting.service.WaitingService;
