+++ src/main/java/roomescape/registration/domain/reservation/controller/ReservationController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.annotation.LoginMemberId;
 import roomescape.client.payment.dto.TossPaymentConfirmRequest;
-import roomescape.payment.PaymentService;
+import roomescape.client.payment.service.PaymentService;
 import roomescape.registration.domain.reservation.domain.Reservation;
 import roomescape.registration.domain.reservation.dto.ReservationRequest;
 import roomescape.registration.domain.reservation.dto.ReservationResponse;
